Updates the code for a newer gcc
GCC was not happy at all with this old code. Apparently the inline
semantics has changed in the last few years which was avoiding the
code to compile.

There was also a bug on matrix.c that was accessing the matrix out of
bounds. In most tests the range was from 0 to 29, but in one was from
0 to 31. Oops.

Signed-off-by: Peter Senna Tschudin <peter.senna@gmail.com>